Lab4.py
I wrote a lot of extra code at first and didn't realize that I didn't need to actually write a function for this to work. I also made the mistake instead of putting range(1000) I put len(1000) but obviously that is just for string length so then I just put the integer 1000 and that also didn't work. Instead of making two if statements, using the "or" helped to really simplify my code.
Once I figured out what the Fibonacci sequence was, I was able to duplicate that and only add the even terms that I needed with and if statement. See comments.
Both of these codes word correctly as I tested them in Project Euler.